Add debian-9-armhf container
This commit adds a container definition for debian-9-armhf. This is identical
to debian-9-amd64 with the addition of steps that install `/usr/bin/qemu-arm-static`
on an amd64 container and then copies it into the ARM container. This allows
x86 hosts with `/proc/sys/fs/binfmt_misc/qemu-arm` configured to execute the
ARM binaries in addition to the ability to run the container on ARM hosts.