Fix use of github.ref for tagging by Actions
GitHub Actions set the value of `github.ref` to the fully qualifies
reference name: `refs/tags/<what you named the tag>`. This is...
technically correct, but not that helpful as there is no easy
way to trim off the leading junk when creating environment variables
via an Action Expression.

This commit updates the steps that use the ref value to run sanitize
it via a call to `basename`.