Add GitHub workflow for building raspbian-10-armhf
This build is a bit different than the others as there isn't a good
published Raspbian 10 base image to start from. The result is that
this build is a two-parter. The first generates a Debian 10 container
that has Debuerreotype installed along with the Raspbian GPG keys:

  https://github.com/debuerreotype/debuerreotype

Debuerreotype is the tool used to create the official Debian docker
images and it does have some support for building Raspbian. Running
the container uses qemu-debootstrap to create a sysroot tarball.

The second stage of the build imports the tarball and adds the normal
"Puppet dev" stuff: SSHD and SystemD.