Give Raspbian some time to boot
This commit adds a 5 second sleep between launching the Raspbian 10
container under Litmus and running the tests. This gives some time
for the emulated system to come to a running state.

This could likely be replaced with a plan that runs the
wait_until_available function with a reasonable timeout.