Update docker-run-args for raspbian-11-armhf
Modern Linux OS builds, like Ubuntu 22.04, run with SystemD and
CGroupsv2 in a configuration where mounting `/sys/fs/cgroup` into
a SystemD container as read-only no longer works. This commit
updates `docker-run-args` to handle this case by running the
container with full privileges and merging it with the host
CGroup structure.

This may not work on older Linux hosts that do not use
`systemd.enableUnifiedCgroupHierarchy=true`, but optimizing for
the future seems better than mantaining compatibility with the
past.